pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bira", email = "" }
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]

[project.scripts]
python-template = "python_template:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.6",
]
